🏷️ Adopt variadic tuples for signatures of tuple (#1516)
